Add tmux conf backup/restore

+++ backup.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+echo 'Creating backup folders'
+mkdir -p {nvim,rectangle,zshrc,iterm2,powerlevel10k,git,tmux}
+
 echo 'Backing up all dot files'
 
 echo 'neovim config'
@@ -9,7 +12,11 @@ echo 'rectangle'
 cp ~/Library/Preferences/com.knollsoft.Rectangle.plist ./rectangle/
 
 echo 'zshrc'
-cp ~/.zshrc ./zsh/.zshrc
+if test -f "$HOME/.is_personal_computer"; then
+	cp ~/.zshrc ./zsh/.zshrc
+else
+	echo "Skipping backup of .zshrc"
+fi
 
 echo 'iterm2'
 cp ~/Library/Preferences/com.googlecode.iterm2.plist ./iterm2/
@@ -20,5 +27,8 @@ cp ~/.p10k.zsh ./powerlevel10k/
 echo 'git global config'
 cp ~/.gitconfig ./git/
 cp ~/.gitconfig-personal ./git/
+
+echo 'tmux'
+cp ~/.tmux.conf ./tmux/
 
 echo "all done!"
